fix: batch id type and make it an url entity
Come on, we're at 4th grade, we should know the differences between a
long and Long (and, if we use frameworks that do some Reflection
programming under the hood, we should use allways reference types, not
primitive ones, since some cases don't do the wrapping and unwrapping
by default).

About making it an url entity: As a user of the API I think
that creating a batch from the frontend makes sense.

Co-authored-by: FerranAD <farandomingo@gmail.com>
Co-authored-by: Pablito2020 <pablofrailealonso@gmail.com>